Fixed code errors

+++ DeviceCode/humidity_sens.h
@@ -2,9 +2,10 @@
 #define HUMIDITY_SENS
 
 #include "periph/adc.h"
+#include <stdio.h>
 
 double calculateMoisture(double rawval);
 
-float sample_moisture(adc_t line, adc_res_t res);
+double sample_moisture(adc_t line, adc_res_t res);
 
 #endif
